Temp fix: Rely on the message length in the prefix rather than stream close
