+++ wire/src/main/java/org/ethereum/beacon/wire/impl/libp2p/encoding/SSZMessageCodec.java
@@ -1,7 +1,6 @@
 package org.ethereum.beacon.wire.impl.libp2p.encoding;
 
 import io.netty.buffer.ByteBuf;
-import io.netty.handler.codec.CorruptedFrameException;
 import org.ethereum.beacon.ssz.SSZSerializer;
 import org.ethereum.beacon.wire.exceptions.WireRpcMalformedException;
 import org.ethereum.beacon.wire.exceptions.WireRpcRemoteError;
@@ -45,12 +44,12 @@ public class SSZMessageCodec<TRequest, TResponse> implements RpcMessageCodec<TRe
 
   private void serializeMsg(Object msg, ByteBuf buf) {
     byte[] msgBytes = sszSerializer.encode(msg);
-    writeRawVarint32(buf, msgBytes.length);
+    Util.writeRawVarint32(buf, msgBytes.length);
     buf.writeBytes(msgBytes);
   }
 
   private <TMessage> TMessage deserializeMsg(ByteBuf buf, Class<TMessage> clazz) {
-    int msgSize = readRawVarint32(buf);
+    int msgSize = Util.readRawVarint32(buf);
     if (msgSize != buf.readableBytes()) {
       throw new WireRpcMalformedException("Size in header (" + msgSize + ") doesn't match payload size: " + buf.readableBytes());
     }
@@ -77,7 +76,7 @@ public class SSZMessageCodec<TRequest, TResponse> implements RpcMessageCodec<TRe
       if (msg.getValue1() != null) {
         buf.writeByte(ErrorCode.ServerError.getCode());
         byte[] errMsgBytes = msg.getValue1().getMessage().getBytes();
-        writeRawVarint32(buf, errMsgBytes.length);
+        Util.writeRawVarint32(buf, errMsgBytes.length);
         buf.writeBytes(errMsgBytes);
       } else {
         buf.writeByte(ErrorCode.OK.getCode());
@@ -94,7 +93,7 @@ public class SSZMessageCodec<TRequest, TResponse> implements RpcMessageCodec<TRe
       if (error == ErrorCode.OK) {
         message = deserializeMsg(buf, responseClass);
       } else {
-        int msgSize = readRawVarint32(buf);
+        int msgSize = Util.readRawVarint32(buf);
         if (msgSize != buf.readableBytes()) {
           throw new WireRpcMalformedException("Size in header (" + msgSize + ") doesn't match payload size: " + buf.readableBytes());
         }
@@ -107,72 +106,4 @@ public class SSZMessageCodec<TRequest, TResponse> implements RpcMessageCodec<TRe
     }
   }
 
-  /**
-   * Encodes int as Protobuf varint
-   * Copied from https://github.com/netty/netty/blob/00afb19d7a37de21b35ce4f6cb3fa7f74809f2ab/codec/src/main/java/io/netty/handler/codec/protobuf/ProtobufVarint32LengthFieldPrepender.java#L58
-   */
-  static void writeRawVarint32(ByteBuf out, int value) {
-    while (true) {
-      if ((value & ~0x7F) == 0) {
-        out.writeByte(value);
-        return;
-      } else {
-        out.writeByte((value & 0x7F) | 0x80);
-        value >>>= 7;
-      }
-    }
-  }
-
-  /**
-   * Decodes Protobuf varint
-   * Copied from: https://github.com/netty/netty/blob/00afb19d7a37de21b35ce4f6cb3fa7f74809f2ab/codec/src/main/java/io/netty/handler/codec/protobuf/ProtobufVarint32FrameDecoder.java#L73
-   */
-  public static int readRawVarint32(ByteBuf buffer) {
-    if (!buffer.isReadable()) {
-      return 0;
-    }
-    buffer.markReaderIndex();
-    byte tmp = buffer.readByte();
-    if (tmp >= 0) {
-      return tmp;
-    } else {
-      int result = tmp & 127;
-      if (!buffer.isReadable()) {
-        buffer.resetReaderIndex();
-        return 0;
-      }
-      if ((tmp = buffer.readByte()) >= 0) {
-        result |= tmp << 7;
-      } else {
-        result |= (tmp & 127) << 7;
-        if (!buffer.isReadable()) {
-          buffer.resetReaderIndex();
-          return 0;
-        }
-        if ((tmp = buffer.readByte()) >= 0) {
-          result |= tmp << 14;
-        } else {
-          result |= (tmp & 127) << 14;
-          if (!buffer.isReadable()) {
-            buffer.resetReaderIndex();
-            return 0;
-          }
-          if ((tmp = buffer.readByte()) >= 0) {
-            result |= tmp << 21;
-          } else {
-            result |= (tmp & 127) << 21;
-            if (!buffer.isReadable()) {
-              buffer.resetReaderIndex();
-              return 0;
-            }
-            result |= (tmp = buffer.readByte()) << 28;
-            if (tmp < 0) {
-              throw new CorruptedFrameException("malformed varint.");
-            }
-          }
-        }
-      }
-      return result;
-    }
-  }
 }
